+++ electron/preload.cjs
@@ -1,0 +1,5 @@
+const { contextBridge, ipcRenderer } = require('electron');
+
+contextBridge.exposeInMainWorld('electronAPI', {
+  writeLog: (logEntry) => ipcRenderer.send('write-log', logEntry)
+});